requirements.txt

# first upgrade your pip:-           pip install --upgrade pip



#New updated version of all requirements
# Core
numpy==1.26.4
pandas==2.2.3
python-dotenv==1.0.1
tqdm==4.66.5
requests==2.32.5

# PDF Processing
PyMuPDF==1.24.10
pypdf==4.3.1
pdfminer.six==20231228

# Embeddings + Vector DB
sentence-transformers==3.1.1
faiss-cpu==1.13.1
scikit-learn==1.5.2

# LLM frameworks
transformers==4.41.2
accelerate==0.30.1
huggingface-hub==0.36.0

# Torch (CPU only to avoid 800MB files)
torch==2.3.0+cpu
# installation command:
# pip install torch --index-url https://download.pytorch.org/whl/cpu

# Gemini API
google-generativeai==0.8.3
google-api-core==2.28.1
google-auth==2.43.0

# FastAPI (optional API use)
fastapi==0.115.0
uvicorn==0.30.3
python-multipart==0.0.9

